style: big refactoring of instructions
Now we model instructions and their arguments in a more "rust-y" way,
using a proper enum and resolving the arguments when we parse the
instruction, and not later when we try to execute it.

This also means that we can handle things like `goto` offset in a much
cleaner way, resolving them when we read the bytecode.