feat: huge refactoring of memory storage!
We have gotten rid of the distinction between `Array` and `Object` -
now we have an `AbstractObject` in memory which can represent both.
This will greatly simplify the GC, even if it introduced some annoyance
in the existing code. Need to do some more refactorings, but this
commmit is already messy enough.